+++ src/cenreg/distribution/cdf.py
@@ -1,3 +1,4 @@
+import warnings
 from typing import Literal
 
 import numpy as np
@@ -44,38 +45,82 @@ class CumulativeDist:
         """
 
         b = np.array(b)
-        assert len(b.shape) == 1
+        if b.ndim != 1:
+            raise ValueError("b must be a one-dimensional array.")
+        if not np.issubdtype(b.dtype, np.floating):
+            warnings.warn("b is not float, converting to float.", stacklevel=2)
+            b = b.astype(float)
+        self.b = b
+
         if cum_p is not None:
+            if p is not None:
+                raise ValueError("Only one of p and cum_p can be provided.")
             cum_p = np.array(cum_p)
-            assert len(cum_p.shape) <= 2
-            if len(cum_p.shape) == 1:
-                assert b.shape[0] - 1 == cum_p.shape[0]
-                assert np.all(np.diff(cum_p) >= 0.0), "cum_p must be non-decreasing"
-            else:
-                assert b.shape[0] - 1 == cum_p.shape[1]
-                assert np.all(np.diff(cum_p, axis=1) >= 0.0), "cum_p must be non-decreasing"
+            self._validate_and_set_cum_p(cum_p, b)
         elif p is not None:
             p = np.array(p)
-            assert len(p.shape) <= 2
-            assert np.all(p >= 0.0)
-            if p.ndim == 1:
-                assert b.shape[0] - 1 == p.shape[0]
-                cum_p = np.cumsum(p)
-            else:
-                assert b.shape[0] - 1 == p.shape[1]
-                cum_p = np.cumsum(p, axis=1)
+            self._validate_p_and_set_cum_p(p, b)
         else:
             raise ValueError("Either p or cum_p must be provided.")
-        assert interpolate in ["linear", "left", "right"]
-        if confidence_interval is not None:
-            assert confidence_interval.shape == (len(cum_p), 2)
 
-        self.b = b
-        self.cum_p = cum_p
+        if interpolate not in ["linear", "left", "right"]:
+            raise ValueError("interpolate must be 'linear', 'left', or 'right'.")
         self.interpolate = interpolate
+
+        if confidence_interval is not None:
+            self._validate_confidence_interval(confidence_interval, self.cum_p)
         self.confidence_interval = confidence_interval
 
-    def cdf(self, y: int | float | np.ndarray):
+    def _validate_and_set_cum_p(self, cum_p: np.ndarray, b: np.ndarray):
+        if cum_p.ndim > 2:
+            raise ValueError("cum_p must be one-dimensional or two-dimensional.")
+        if cum_p.ndim == 1:
+            if b.shape[0] - 1 != cum_p.shape[0]:
+                raise ValueError("Length of cum_p must be one less than length of b.")
+            if np.any(cum_p < 0.0) or np.any(cum_p > 1.0):
+                raise ValueError("cum_p must be in the range [0.0, 1.0].")
+            if np.all(np.diff(cum_p) < 0.0):
+                raise ValueError("cum_p must be non-decreasing.")
+        else:  # cum_p.ndim == 2
+            if b.shape[0] - 1 != cum_p.shape[1]:
+                raise ValueError("Length of cum_p must be one less than length of b.")
+            if np.any(np.diff(cum_p, axis=1) < 0.0):  # allow cum_p[i, j] == cum_p[i, j + 1]
+                raise ValueError("cum_p must be non-decreasing.")
+        # check if cum_p is float
+        if not np.issubdtype(cum_p.dtype, np.floating):
+            warnings.warn("cum_p is not float, converting to float.", stacklevel=2)
+            cum_p = cum_p.astype(float)
+        self.cum_p = cum_p
+
+    def _validate_p_and_set_cum_p(self, p: np.ndarray, b: np.ndarray):
+        if p.ndim > 2:
+            raise ValueError("p must be one-dimensional or two-dimensional.")
+        if np.any(p < 0.0):
+            raise ValueError("p must be non-negative.")
+        if not np.issubdtype(p.dtype, np.floating):
+            warnings.warn("p is not float, converting to float.", stacklevel=2)
+            p = p.astype(float)
+        if p.ndim == 1:
+            if b.shape[0] - 1 != p.shape[0]:
+                raise ValueError("Length of p must be one less than length of b.")
+            cum_p = np.cumsum(p)
+        else:  # p.ndim == 2
+            if b.shape[0] - 1 != p.shape[1]:
+                raise ValueError("Length of p must be one less than length of b.")
+            cum_p = np.cumsum(p, axis=1)
+        self.cum_p = cum_p
+
+    def _validate_confidence_interval(self, confidence_interval: np.ndarray, cum_p: np.ndarray):
+        if confidence_interval.shape[0] != cum_p.shape[0]:
+            raise ValueError("confidence_interval must have the same number of rows as cum_p.")
+        if confidence_interval.shape[1] != 2:
+            raise ValueError("confidence_interval must have two columns.")
+        if np.any(confidence_interval < 0.0) or np.any(confidence_interval > 1.0):
+            raise ValueError("confidence_interval must be in the range [0.0, 1.0].")
+        if np.any(confidence_interval[:, 0] > confidence_interval[:, 1]):
+            raise ValueError("Lower bound of confidence interval must be less than or equal to upper bound.")
+
+    def cdf(self, y: float | np.ndarray) -> np.ndarray:
         """
         Cumulative distribution function (i.e., inverse of quantile function).
 
@@ -83,6 +128,7 @@ class CumulativeDist:
         -------
         y : np.ndarray | float
             Values for which the CDF is computed.
+            y can be a scalar or a one-dimensional array or a two-dimensional array.
 
         Returns
         -------
@@ -92,19 +138,24 @@ class CumulativeDist:
         if isinstance(y, int | float):
             y = np.array([y], dtype=float)
 
-        if self.cum_p.ndim == 1:
-            assert y.ndim == 1
-        elif self.cum_p.ndim == 2:
-            assert y.ndim == 1 or y.ndim == 2
+        # check input y
+        if isinstance(y, np.ndarray):
+            if len(y.shape) > 2:
+                raise ValueError("y must be a scalar or a one-dimensional array or a two-dimensional array.")
         else:
-            raise ValueError("cum_p must be one-dimensional or two-dimensional.")
-
+            if isinstance(y, float):
+                y = np.array([y])
+            else:
+                raise ValueError("y must be a scalar or a one-dimensional array or a two-dimensional array.")
+        if not np.issubdtype(y.dtype, np.floating):
+            warnings.warn("y is not float, converting to float.", stacklevel=2)
+            y = y.astype(float)
         if self.cum_p.ndim == 2 and y.ndim == 1:
             y = np.tile(y, (self.cum_p.shape[0], 1))
 
         if self.interpolate == "linear":
             # linear interpolation implementation
-            ret = np.zeros_like(y)
+            ret = np.zeros_like(y, dtype=float)
             mask_low = y <= self.b[0]
             ret[mask_low] = 0.0
             mask_high = y > self.b[-1]
@@ -117,8 +168,7 @@ class CumulativeDist:
                 zeros = np.zeros((self.cum_p.shape[0], 1))
                 temp = np.concatenate((zeros, self.cum_p), axis=1)
                 temp2 = linear_interpolation(self.b, temp, y)
-                ret[mask] = temp2[mask]
-                # ret[mask] = linear_interpolation(boundaries, temp, y)[mask]
+                ret[mask] = temp2[mask].flatten()
             return ret
         else:
             # step function implementation
@@ -156,12 +206,7 @@ class CumulativeDist:
             Compute inverse CDF values for each value in quantiles.
         """
 
-        if isinstance(quantiles, int | float):
-            quantiles = np.array([quantiles], dtype=float)
-        if np.any(quantiles < 0.0):
-            raise ValueError("quantiles must be non-negative.")
-        if np.any(quantiles > 1.0):
-            raise ValueError("quantiles must be at most 1.0.")
+        quantiles = self._validate_quantiles(quantiles)
         if self.cum_p.ndim == 2 and quantiles.ndim == 1:
             quantiles = np.tile(quantiles, (self.cum_p.shape[0], 1))
 
@@ -204,6 +249,18 @@ class CumulativeDist:
             mask = ~(mask_low | mask_high)
             ret[mask] = self.b[idx[mask]]
             return ret
+
+    def _validate_quantiles(self, quantiles):
+        if isinstance(quantiles, float):
+            quantiles = np.array([quantiles])
+        elif not np.issubdtype(quantiles.dtype, np.floating):
+            warnings.warn("quantiles is not float, converting to float.", stacklevel=2)
+            quantiles = quantiles.astype(float)
+        if np.any(quantiles < 0.0):
+            raise ValueError("quantiles must be non-negative.")
+        if np.any(quantiles > 1.0):
+            raise ValueError("quantiles must be at most 1.0.")
+        return quantiles
 
     def survival_function(self, y: float | np.ndarray) -> np.ndarray:
         """