+++ src/cenreg/metric/cdf.py
@@ -1,8 +1,8 @@
 import numpy as np
 
-import cenreg.model.kaplan_meier as kaplan_meier
+from cenreg.model.nonparametric import kaplan_meier_estimator
 
-
+'''
 def negative_loglikelihood(
     dist,
     observed_times: np.ndarray,
@@ -72,7 +72,182 @@ def negative_loglikelihood(
     last_idx = idx == len(boundaries) - 2
     loss[~uncensored & ~last_idx] = -np.log(1.0 - F_ub[~uncensored & ~last_idx] + eps)
     loss[~uncensored & last_idx] = -np.log(1.0 - F_lb[~uncensored & last_idx] + eps)
-    return loss
+    return loss.sum()
+
+
+def negative_log_likelihood_interval(
+    dist,
+    lb: np.ndarray,
+    ub: np.ndarray,
+    eps: float = 0.0001,
+) -> np.ndarray:
+    """
+    Compute Negative log-likelihood for interval-censored data.
+
+    Parameters
+    ----------
+    dist: distribution object
+        Prediction results.
+    lb: ndarray
+        Lower bounds of the intervals.
+    ub: ndarray
+        Upper bounds of the intervals.
+    eps: float
+        Small positive value for numerical stability.
+
+    Returns
+    -------
+    loss : ndarray
+        Value of negative log-likelihood.
+    """
+
+    if np.any(lb >= ub):
+        raise ValueError("lb must be strictly smaller than ub for all data points.")
+
+    mask = (dist.b >= lb.reshape(-1, 1)) & (dist.b < ub.reshape(-1, 1))
+    p = np.diff(dist.cum_p, axis=0)
+    loss = -np.log(np.sum(p * mask.astype(np.float32), axis=1) + eps)
+    return loss.sum()
+'''
+
+
+def _negative_log_likelihood(
+    dist,
+    lb: np.ndarray,
+    ub: np.ndarray,
+    proportional: bool = True,
+    eps: float = 0.0001,
+) -> np.ndarray:
+    """
+    Compute Negative log-likelihood.
+
+    Parameters
+    ----------
+    dist: predicted distribution
+
+    lb: ndarray of shape [batch_size]
+        lower bound of the interval-censored data
+
+    ub: ndarray of shape [batch_size]
+        upper bound of the interval-censored data
+
+    proportional: bool
+        whether to distribute the probability mass proportionally for censored data
+
+    eps: float
+        small value to avoid numerical issues
+
+    Returns
+    -------
+    loss : ndarray of shape [batch_size]
+    """
+
+    y_bins = dist.b
+    idx_lb = np.searchsorted(y_bins, lb.reshape(-1, 1), side="right")
+    idx_lb = np.clip(idx_lb, 1, len(y_bins) - 1)
+    idx_ub = np.searchsorted(y_bins, ub.reshape(-1, 1), side="left")
+    idx_ub = np.clip(idx_ub, 1, len(y_bins) - 1)
+    b_lb = y_bins[idx_lb - 1]
+    b_ub = y_bins[idx_ub]
+    F_lb = dist.cdf(b_lb)
+    F_ub = dist.cdf(b_ub)
+
+    if proportional:
+        interval = (idx_lb + 1 == idx_ub)[:, 0]
+        F_lb[interval] = dist.cdf(lb.reshape(-1, 1))[interval]
+        F_ub[interval] = dist.cdf(ub.reshape(-1, 1))[interval]
+
+    loss = -np.log(F_ub - F_lb + eps)
+    return loss.reshape(-1)
+
+
+def negative_log_likelihood_survival(
+    dist,
+    y: np.ndarray,
+    uncensored: np.ndarray | None = None,
+    proportional: bool = True,
+    eps: float = 0.0001,
+) -> np.ndarray:
+    """
+    Compute Negative log-likelihood for survival data.
+
+    Parameters
+    ----------
+    dist: predicted distribution
+        Must be an instance of cenreg.distribution.cdf.CumulativeDist with linear interpolation.
+
+    y: ndarray of shape [batch_size]
+        observed times
+
+    uncensored: ndarray of shape [batch_size]
+        boolean array indicating whether the event was observed (True) or censored (False)
+
+    proportional: bool
+        whether to distribute the probability mass proportionally for censored data
+
+    eps: float
+        small value to avoid numerical issues
+
+    Returns
+    -------
+    loss : ndarray of shape [batch_size]
+    """
+
+    if len(y.shape) != 1:
+        raise ValueError("y must be of shape [batch_size]")
+    if uncensored is None:
+        uncensored = np.ones(y.shape, dtype=bool)
+    if dist.interpolate != "linear":
+        raise Warning("dist must be an instance of cenreg.distribution.cdf.CumulativeDist with linear interpolation.")
+
+    lb = y.astype(float)
+    ub = y.astype(float)
+    ub[~uncensored] = np.inf
+    return _negative_log_likelihood(dist, lb, ub, proportional, eps)
+
+
+def negative_log_likelihood_interval(
+    dist,
+    lb: np.ndarray,
+    ub: np.ndarray,
+    proportional: bool = True,
+    eps: float = 0.0001,
+) -> np.ndarray:
+    """
+    Compute Negative log-likelihood for interval-censored data.
+
+    Parameters
+    ----------
+    dist: predicted distribution
+        Must be an instance of cenreg.distribution.cdf.CumulativeDist with linear interpolation.
+
+    lb: ndarray of shape [batch_size]
+        lower bounds of the intervals
+
+    ub: ndarray of shape [batch_size]
+        upper bounds of the intervals
+
+    proportional: bool
+        whether to distribute the probability mass proportionally for censored data
+
+    eps: float
+        small value to avoid numerical issues
+
+    Returns
+    -------
+    loss : ndarray of shape [batch_size]
+    """
+
+    if len(lb.shape) != 1:
+        raise ValueError("lb must be of shape [batch_size]")
+    if len(ub.shape) != 1:
+        raise ValueError("ub must be of shape [batch_size]")
+    if len(lb) != len(ub):
+        raise ValueError("lb and ub must have the same length")
+    if dist.interpolate != "linear":
+        raise Warning("dist must be an instance of cenreg.distribution.cdf.CumulativeDist with linear interpolation.")
+
+    return _negative_log_likelihood(dist, lb, ub, proportional, eps)
 
 
 def brier(
@@ -212,7 +387,7 @@ def km_calibration(
     y_bins: ndarray
         Bins for the prediction to be evaluated.
         Each element in observed_times must be STRICTLY smaller than y_bins[-1]
-    EPS : float
+    eps : float
         Small positive value for numerical stability
 
     Returns
@@ -228,14 +403,13 @@ def km_calibration(
         uncensored = uncensored.astype(bool)
     if y_bins is None:
         try:
-            y_bins = dist.get_boundaries()
+            y_bins = dist.b
         except AttributeError as err:
-            raise ValueError("y_bins must be provided if pred does not have y_bins.") from err
+            raise ValueError("y_bins must be provided if dist does not have attribute 'b'.") from err
 
     # compute Kaplan-Meier distribution and prediction distribution
-    km = kaplan_meier.KaplanMeierDistribution()
-    km.fit(observed_times, uncensored)
-    last_idx = np.searchsorted(y_bins, km.last_uncensored_time, side="right") - 1
+    km = kaplan_meier_estimator(observed_times, uncensored)
+    last_idx = np.searchsorted(y_bins, km.b[-1], side="right") - 1
     F_km = km.cdf(y_bins)
     F_km[-1] = 1.0
     f_km = F_km[1:] - F_km[:-1]