+++ src/cenreg/pytorch/loss_cdf.py
@@ -1,14 +1,15 @@
+import numpy as np
 import torch
 import torch.nn.functional as F
 
-from cenreg.pytorch.distribution import LinearCDF
+from cenreg.pytorch.distribution import CumulativeDist, LinearCDF
 
 
 def negative_log_likelihood(
     dist,
-    y: torch.Tensor,
-    y_bins: torch.Tensor | None = None,
-    uncensored: torch.Tensor | None = None,
+    lb: torch.Tensor,
+    ub: torch.Tensor,
+    proportional: bool = True,
     eps: float = 0.0001,
 ) -> torch.Tensor:
     """
@@ -18,64 +19,55 @@ def negative_log_likelihood(
     ----------
     dist: predicted distribution
 
-    y: Tensor of shape [batch_size, 1]
+    lb: Tensor of shape [batch_size]
+        lower bound of the interval-censored data
 
-    y_bins: Tensor of shape [num_bins+1]
+    ub: Tensor of shape [batch_size]
+        upper bound of the interval-censored data
 
-    uncensored: Tensor of shape [batch_size]
+    proportional: bool
+        whether to distribute the probability mass proportionally for censored data
 
     eps: float
+        small value to avoid numerical issues
 
     Returns
     -------
     loss : Tensor of shape [batch_size]
     """
 
-    if y_bins is None:
-        y_bins = dist.boundaries
-    if uncensored is not None:
-        uncensored = uncensored.bool()
+    y_bins = dist.b
+    idx_lb = torch.searchsorted(y_bins, lb.view(-1, 1), side="right")
+    idx_lb = torch.clamp(idx_lb, min=1, max=len(y_bins) - 1)
+    idx_ub = torch.searchsorted(y_bins, ub.view(-1, 1), side="left")
+    idx_ub = torch.clamp(idx_ub, min=1, max=len(y_bins) - 1)
+    b_lb = y_bins[idx_lb - 1]
+    b_ub = y_bins[idx_ub]
+    F_lb = dist.cdf(b_lb)
+    F_ub = dist.cdf(b_ub)
 
-    idx = torch.searchsorted(y_bins, y.view(-1, 1), right=True)
-    b_lb = y_bins[idx - 1]
-    b_ub = y_bins[idx]
-    F_lb = dist.cdf(b_lb.view(-1, 1))
-    F_ub = dist.cdf(b_ub.view(-1, 1))
-    if uncensored is None:
-        F_lb_uncensored = F_lb
-        F_ub_uncensored = F_ub
-    else:
-        uncensored = uncensored.bool()
-        F_lb_uncensored = F_lb[uncensored]
-        F_ub_uncensored = F_ub[uncensored]
+    if proportional:
+        interval = (idx_lb + 1 == idx_ub)[:, 0]
+        F_lb[interval] = dist.cdf(lb.view(-1, 1))[interval]
+        F_ub[interval] = dist.cdf(ub.view(-1, 1))[interval]
 
-    loss = torch.zeros(y.shape[0], 1, device=y.device)
-    pu = F_ub_uncensored - F_lb_uncensored + eps
-    loss[uncensored] = -torch.log(pu)
-    if uncensored is not None:
-        F_lb_censored = F_lb[~uncensored]
-        F_ub_censored = F_ub[~uncensored]
-        c = y[~uncensored].view(-1, 1)
-        F_c = dist.cdf(c, ~uncensored)
-        denominator = torch.clamp(1.0 - F_c, min=eps)
-        w = torch.clamp((F_ub_censored - F_c) / denominator, min=0.0, max=1.0)
-        w = w.detach()
-        pc1 = F_ub_censored - F_lb_censored + eps
-        loss[~uncensored, :] -= w * torch.log(pc1)
-        pc2 = 1.0 - F_ub_censored + eps
-        loss[~uncensored, :] -= (1.0 - w) * torch.log(pc2)
-    return loss
+    loss = -torch.log(F_ub - F_lb + eps)
+    return loss.view(-1)
 
 
-class NegativeLogLikelihood:
+class NegativeLogLikelihoodSurvival:
     """
-    Loss class for negative log-likelihood.
+    Loss class for negative log-likelihood for right-censored data.
     """
 
-    def __init__(self, y_bins: torch.Tensor, apply_cumsum: bool = True):
-        self.distribution = LinearCDF(y_bins)
+    def __init__(self, y_bins: torch.Tensor, apply_cumsum: bool = True, proportional: bool = True):
+        if not isinstance(y_bins, torch.Tensor):
+            raise ValueError("y_bins should be a torch.Tensor")
+
+        self.distribution = CumulativeDist(y_bins)
         self.y_bins = y_bins
         self.apply_cumsum = apply_cumsum
+        self.proportional = proportional
         self.eps = 0.0001
 
     def loss(
@@ -84,11 +76,74 @@ class NegativeLogLikelihood:
         y: torch.Tensor,
         uncensored: torch.Tensor | None = None,
     ) -> torch.Tensor:
-        assert len(pred.shape) == 2
-        assert len(y.shape) == 1
+        if len(pred.shape) != 2:
+            raise ValueError("pred must be of shape [batch_size, num_bins]")
+        if len(y.shape) != 1:
+            raise ValueError("y must be of shape [batch_size]")
+        if len(pred) != len(y):
+            raise ValueError("pred and y must have the same length")
+        if uncensored is None:
+            raise ValueError("uncensored must be provided")
 
         self.distribution.set_knot_values(pred, apply_cumsum=self.apply_cumsum)
-        return negative_log_likelihood(self.distribution, y, self.y_bins, uncensored, self.eps)
+        lb = y
+        ub = y
+        ub[~uncensored] = np.inf
+        return negative_log_likelihood(self.distribution, lb, ub, self.proportional, self.eps)
+
+
+class NegativeLogLikelihoodInterval:
+    """
+    Loss class for negative log-likelihood for interval-censored data.
+    """
+
+    def __init__(
+        self,
+        y_bins: torch.Tensor,
+        apply_cumsum: bool = True,
+        proportional: bool = True,
+        eps: float = 0.0001,
+    ):
+        if not isinstance(y_bins, torch.Tensor):
+            raise ValueError("y_bins should be a torch.Tensor")
+        if len(y_bins.shape) != 1:
+            raise ValueError("y_bins should be a 1D tensor")
+        diff = torch.diff(y_bins)
+        if torch.any(diff <= 0.0):
+            raise ValueError("y_bins should be sorted in ascending order")
+        if y_bins[0] == -np.inf:
+            raise ValueError("y_bins should not contain -inf")
+        if y_bins[-1] == np.inf:
+            raise ValueError("y_bins should not contain inf")
+
+        self.distribution = CumulativeDist(y_bins)
+        self.y_bins = y_bins
+        self.apply_cumsum = apply_cumsum
+        self.proportional = proportional
+        self.eps = eps
+
+    def loss(
+        self,
+        pred: torch.Tensor,
+        lb: torch.Tensor,
+        ub: torch.Tensor,
+    ) -> torch.Tensor:
+        if len(pred.shape) != 2:
+            raise ValueError("pred must be of shape [batch_size, num_bins]")
+        if len(lb.shape) != 1:
+            raise ValueError("lb must be of shape [batch_size]")
+        if len(ub.shape) != 1:
+            raise ValueError("ub must be of shape [batch_size]")
+        if len(lb) != len(ub):
+            raise ValueError("lb and ub must have the same length")
+        if len(pred) != len(lb):
+            raise ValueError("pred and lb must have the same length")
+
+        if self.apply_cumsum:
+            self.distribution.set_knot_values(p=pred)
+        else:
+            self.distribution.set_knot_values(cum_p=pred)
+        return negative_log_likelihood(self.distribution, lb, ub, self.proportional, self.eps)
 
 
 class CNLLCR:
@@ -97,6 +152,9 @@ class CNLLCR:
     """
 
     def __init__(self, boundaries: torch.Tensor, num_risks: int):
+        if not isinstance(boundaries, torch.Tensor):
+            raise ValueError("boundaries should be a torch.Tensor")
+
         self.max_time = boundaries[-1]
         self.list_distribution = []
         for _ in range(num_risks):
@@ -178,6 +236,9 @@ class Brier:
     """
 
     def __init__(self, y_bins: torch.Tensor, apply_cumsum: bool = True):
+        if not isinstance(y_bins, torch.Tensor):
+            raise ValueError("y_bins should be a torch.Tensor")
+
         self.distribution = LinearCDF(y_bins)
         self.y_bins = y_bins
         self.apply_cumsum = apply_cumsum
@@ -235,6 +296,9 @@ class RankedProbabilityScore:
     """
 
     def __init__(self, y_bins: torch.Tensor, apply_cumsum: bool = True):
+        if not isinstance(y_bins, torch.Tensor):
+            raise ValueError("y_bins should be a torch.Tensor")
+
         self.distribution = LinearCDF(y_bins)
         self.y_bins = y_bins
         self.apply_cumsum = apply_cumsum