+++ tests/distribution/test_cdf.py
@@ -64,6 +64,20 @@ class TestCumulativeDist(unittest.TestCase):
         self.assertAlmostEqual(ret[3].item(), 2.0)
         self.assertAlmostEqual(ret[4].item(), 2.0)
 
+    def test_left3(self):
+        omega = np.array([-1.0, 0.0, 1.0, 2.0, 3.0, 4.0])
+        cum_p = np.array([0.0, 0.15, 0.55, 0.85, 1.0])
+        dist = CumulativeDist(b=omega, cum_p=cum_p, interpolate="left")
+
+        ret = dist.cdf(np.array([-1.0, 0.0, 1.0, 2.0, 3.0, 4.0]))
+        self.assertEqual(ret.shape, (6,))
+        self.assertAlmostEqual(ret[0].item(), 0.0)
+        self.assertAlmostEqual(ret[1].item(), 0.0)
+        self.assertAlmostEqual(ret[2].item(), 0.15)
+        self.assertAlmostEqual(ret[3].item(), 0.55)
+        self.assertAlmostEqual(ret[4].item(), 0.85)
+        self.assertAlmostEqual(ret[5].item(), 1.0)
+
     def test_right1(self):
         dist = CumulativeDist(
             b=np.array([0.0, 1.0, 2.0, 3.0, 4.0]),
@@ -121,6 +135,26 @@ class TestCumulativeDist(unittest.TestCase):
         self.assertAlmostEqual(ret[2].item(), 1.0)
         self.assertAlmostEqual(ret[3].item(), 2.0)
         self.assertAlmostEqual(ret[4].item(), 2.0)
+
+    def test_right3(self):
+        dist = CumulativeDist(
+            b=np.array([0.0, 1.0, 2.0, 3.0]),
+            cum_p=np.array([0.0, 0.3, 1.0]),
+            interpolate="right",
+        )
+
+        ret = dist.cdf(np.array([[-0.5, 0.0, 0.5, 1.0, 1.5], [1.5, 2.0, 2.5, 3.0, 3.5]]))
+        self.assertEqual(ret.shape, (2, 5))
+        self.assertAlmostEqual(ret[0, 0].item(), 0.0)
+        self.assertAlmostEqual(ret[0, 1].item(), 0.0)
+        self.assertAlmostEqual(ret[0, 2].item(), 0.0)
+        self.assertAlmostEqual(ret[0, 3].item(), 0.3)
+        self.assertAlmostEqual(ret[0, 4].item(), 0.3)
+        self.assertAlmostEqual(ret[1, 0].item(), 0.3)
+        self.assertAlmostEqual(ret[1, 1].item(), 1.0)
+        self.assertAlmostEqual(ret[1, 2].item(), 1.0)
+        self.assertAlmostEqual(ret[1, 3].item(), 1.0)
+        self.assertAlmostEqual(ret[1, 4].item(), 1.0)
 
     def test_linear1(self):
         dist = CumulativeDist(
@@ -187,7 +221,6 @@ class TestCumulativeDist(unittest.TestCase):
             interpolate="linear",
         )
 
-        print("test_linear3")
         ret = dist.cdf(np.array([2.0, 2.5, 3.0]))
         self.assertEqual(ret.shape, (2, 3))
         self.assertAlmostEqual(ret[0, 0].item(), 0.3)
@@ -211,7 +244,6 @@ class TestCumulativeDist(unittest.TestCase):
             interpolate="linear",
         )
 
-        print("test_linear4")
         ret = dist.cdf(np.array([[2.0], [2.5]]))
         self.assertEqual(ret.shape, (2, 1))
         self.assertAlmostEqual(ret[0, 0].item(), 0.3)