+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [0.1.1] - 2026-07-03
+### Added
+- Tutorials for Li-Watkins-Yu estimator and NN-IC-Log model
+- New calibration metric (ic-calibration)
+### Changed
+- APIs for negative log likelihood
+### Removed
+- Old distribution classes (CopulaGraphicDistribution, KaplanMeierDistribution)
+
 ## [0.1.0] - 2026-03-27
 ### Added
 - Tutorial for Turnbull estimator